requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
numpy>=1.26.0
scikit-learn==1.3.2
matplotlib==3.8.2
plotly==5.17.0
textblob==0.17.1
snscrape==0.7.0.20230622
pyarrow==14.0.1
APScheduler==3.10.4
gunicorn==21.2.0
python-dotenv==1.0.0
memory-profiler==0.61.0
psutil==5.9.6
requests==2.31.0
Werkzeug==3.0.1